Make DESTDIR operation match common practice
Modify DESTDIR to match common practice (for instance, that of automake),
in that DESTDIR represents a staged install directory. It is intended
that the tree would need to than be copied to the final install location
afterwards. When supplied, DESTDIR will not change any compiled-in
paths in darshan.

See the automake manual section about DESTDIR for an example of standard
usage.

https://www.gnu.org/software/automake/manual/html_node/DESTDIR.html