check lustre striping for all files
This attempts to work around issue #302, where Darshan can't
instrument Lustre striping information for symlinks to Lustre
files that reside on other file systems. The method for querying
Lustre striping info is general enough (based on xattrs) and
low-overhead enough that this should not create any issues for
Darshan users.

Closes #302