drop '--build-ext' option in setup.py for env var
This option to build the C extension is only used in the wheel
building process. cibuildwheel does not appear to support an
ability to pass options to setuptools, and even if it did there
appears to be no support for pip installing wheels with flags
like '--build-option' (see
https://github.com/pypa/pip/issues/5771).

This commit also updates our build-wheels script to use this
env var, even though those scripts are more or less deprecated.